votiface_django/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
face_recognition==1.3.0
numpy==1.22.1
opencv_python==4.5.5.62
Pyrebase4==4.5.0
requests==2.27.1